% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Compute AUC from prediction statistics.}
\usage{
auc(tpr, fpr)
}
\arguments{
\item{tpr}{A vector of true positive rates.}

\item{fpr}{A vector, of the same length, of false positive rates.}
}
\value{
The area under the ROC curve.
}
\description{
Compute AUC from prediction statistics.
}

