% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc_by}
\alias{auc_by}
\title{Compute AUC for observations drawn from two classes.}
\usage{
auc_by(measure, classes, goldleft)
}
\arguments{
\item{measure}{A vector of observations.}

\item{classes}{A vector of class labels.}

\item{goldleft}{The value of classes corresponding to the "negative" group, i.e.,
the one which will be on the < side of the classifier.}
}
\value{
The area under the ROC curve.
}
\description{
Compute AUC for observations drawn from two classes.
}

