% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{integrate_finite}
\alias{integrate_finite}
\title{Simple numerical integration using the midpoint rule.
Rectangles are unevenly spaced and are as wide as the gaps
between the x's.}
\usage{
integrate_finite(x, y)
}
\arguments{
\item{x}{A variable.}

\item{y}{f(x).}
}
\value{
The approximate area under f(x).
}
\description{
Simple numerical integration using the midpoint rule.
Rectangles are unevenly spaced and are as wide as the gaps
between the x's.
}

